# app/beiboot_api/openapi.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Get Root", "operationId": "get_root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/clusters/": {"get": {"tags": ["clusters"], "summary": "Cluster List", "operationId": "cluster_list_clusters__get", "parameters": [{"required": false, "schema": {"title": "Page", "minimum": 1.0, "type": "integer", "default": 1}, "name": "page", "in": "query"}, {"required": false, "schema": {"title": "Size", "maximum": 100.0, "minimum": 1.0, "type": "integer", "default": 50}, "name": "size", "in": "query"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Page_ClusterStateResponse_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["clusters"], "summary": "Cluster Create", "operationId": "cluster_create_clusters__post", "parameters": [{"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClusterRequest"}, "example": {"name": "hello", "parameters": [{"name": "K8S_VERSION"}, {"name": "NODE_COUNT", "value": 1}]}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClusterStateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clusters/{cluster_name}": {"get": {"tags": ["clusters"], "summary": "Cluster Info", "operationId": "cluster_info_clusters__cluster_name__get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClusterInfoResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["clusters"], "summary": "Cluster Delete", "operationId": "cluster_delete_clusters__cluster_name__delete", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clusters/{cluster_name}/heartbeat": {"get": {"tags": ["clusters"], "summary": "Cluster State", "operationId": "cluster_state_clusters__cluster_name__heartbeat_get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClusterStateResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clusters/{cluster_name}/kubeconfig": {"get": {"tags": ["clusters"], "summary": "Cluster Kubeconfig", "operationId": "cluster_kubeconfig_clusters__cluster_name__kubeconfig_get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clusters/{cluster_name}/parameters": {"get": {"tags": ["clusters"], "summary": "Cluster Parameters", "operationId": "cluster_parameters_clusters__cluster_name__parameters_get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "Page", "minimum": 1.0, "type": "integer", "default": 1}, "name": "page", "in": "query"}, {"required": false, "schema": {"title": "Size", "maximum": 100.0, "minimum": 1.0, "type": "integer", "default": 50}, "name": "size", "in": "query"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Page_dict_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clusters/{cluster_name}/parameters/{parameter_name}": {"get": {"tags": ["clusters"], "summary": "Cluster Parameter", "operationId": "cluster_parameter_clusters__cluster_name__parameters__parameter_name__get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": true, "schema": {"title": "Parameter Name", "type": "string"}, "name": "parameter_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/clusters/ws/{cluster_name}": {"get": {"tags": ["clusters"], "summary": "Get", "operationId": "get_clusters_ws__cluster_name__get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/connections/{cluster_name}/ghostunnel/": {"get": {"tags": ["connections"], "summary": "Ghostunnel", "operationId": "ghostunnel_connections__cluster_name__ghostunnel__get", "parameters": [{"required": true, "schema": {"title": "Cluster Name", "type": "string"}, "name": "cluster_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GhostunnelResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/configs/": {"get": {"tags": ["configs"], "summary": "Config List", "operationId": "config_list_configs__get", "parameters": [{"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/configs/default/refresh/": {"get": {"tags": ["configs"], "summary": "Config Default Refresh", "operationId": "config_default_refresh_configs_default_refresh__get", "parameters": [{"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/configs/{config_name}/refresh/": {"get": {"tags": ["configs"], "summary": "Config Refresh", "operationId": "config_refresh_configs__config_name__refresh__get", "parameters": [{"required": true, "schema": {"title": "Config Name", "type": "string"}, "name": "config_name", "in": "path"}, {"required": false, "schema": {"title": "X-Forwarded-User", "type": "string"}, "name": "x-forwarded-user", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Groups", "type": "string"}, "name": "x-forwarded-groups", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Email", "type": "string"}, "name": "x-forwarded-email", "in": "header"}, {"required": false, "schema": {"title": "X-Forwarded-Preferred-Username", "type": "string"}, "name": "x-forwarded-preferred-username", "in": "header"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/debug/headers": {"get": {"tags": ["debug"], "summary": "Get Headers", "operationId": "get_headers_debug_headers_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/debug/sentry": {"get": {"tags": ["debug"], "summary": "Trigger Error", "operationId": "trigger_error_debug_sentry_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"BeibootState": {"title": "BeibootState", "enum": ["REQUESTED", "CREATING", "PENDING", "RUNNING", "READY", "TERMINATING", "ERROR"], "description": "An enumeration."}, "ClusterInfoResponse": {"title": "ClusterInfoResponse", "required": ["name", "namespace"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "namespace": {"title": "Namespace", "type": "string"}, "state": {"$ref": "#/components/schemas/BeibootState"}, "parameters": {"title": "Parameters", "type": "array", "items": {"anyOf": [{"$ref": "#/components/schemas/StringParameter"}, {"$ref": "#/components/schemas/IntegerParameter"}]}}}}, "ClusterParameter": {"title": "ClusterParameter", "enum": ["K8S_VERSION", "NODE_COUNT", "LIFETIME", "SESSION_TIMEOUT"], "description": "An enumeration."}, "ClusterRequest": {"title": "ClusterRequest", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "parameters": {"title": "Parameters", "type": "array", "items": {"anyOf": [{"$ref": "#/components/schemas/StringParameter"}, {"$ref": "#/components/schemas/IntegerParameter"}]}}, "ports": {"title": "Ports", "type": "array", "items": {"type": "string"}}, "labels": {"$ref": "#/components/schemas/Labels"}}}, "ClusterStateResponse": {"title": "ClusterStateResponse", "required": ["name"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "state": {"$ref": "#/components/schemas/BeibootState"}}}, "GhostunnelPort": {"title": "GhostunnelPort", "required": ["endpoint", "target"], "type": "object", "properties": {"endpoint": {"title": "Endpoint", "type": "string"}, "target": {"title": "Target", "type": "integer"}}}, "GhostunnelResponse": {"title": "GhostunnelResponse", "required": ["mtls"], "type": "object", "properties": {"mtls": {"title": "Mtls", "type": "object"}, "ports": {"title": "Ports", "type": "array", "items": {"$ref": "#/components/schemas/GhostunnelPort"}}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "IntegerParameter": {"title": "IntegerParameter", "required": ["name"], "type": "object", "properties": {"name": {"$ref": "#/components/schemas/ClusterParameter"}, "value": {"title": "Value", "type": "integer"}}}, "Labels": {"title": "Labels", "type": "object", "properties": {"user": {"title": "User", "type": "string"}}}, "Page_ClusterStateResponse_": {"title": "Page[ClusterStateResponse]", "required": ["items", "total"], "type": "object", "properties": {"items": {"title": "Items", "type": "array", "items": {"$ref": "#/components/schemas/ClusterStateResponse"}}, "total": {"title": "Total", "minimum": 0.0, "type": "integer"}, "page": {"title": "Page", "minimum": 1.0, "type": "integer"}, "size": {"title": "Size", "minimum": 1.0, "type": "integer"}, "pages": {"title": "Pages", "minimum": 0.0, "type": "integer"}}}, "Page_dict_": {"title": "Page[dict]", "required": ["items", "total"], "type": "object", "properties": {"items": {"title": "Items", "type": "array", "items": {"type": "object"}}, "total": {"title": "Total", "minimum": 0.0, "type": "integer"}, "page": {"title": "Page", "minimum": 1.0, "type": "integer"}, "size": {"title": "Size", "minimum": 1.0, "type": "integer"}, "pages": {"title": "Pages", "minimum": 0.0, "type": "integer"}}}, "StringParameter": {"title": "StringParameter", "required": ["name"], "type": "object", "properties": {"name": {"$ref": "#/components/schemas/ClusterParameter"}, "value": {"title": "Value", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}